Apply the references view
